apps/example-application/requirements.txt
anyio==4.7.0
bcrypt==4.2.1
boto3==1.17.80
botocore==1.20.112
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
confluent-kafka==1.6.0
cryptography==44.0.0
docutils==0.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm-spectrum-discover-application-sdk==0.0.22
ibm-watson-machine-learning==1.0.178
ibm_watsonx_ai==1.1.24
idna==2.10
importlib_metadata==8.5.0
jmespath==0.10.0
lomond==0.3.3
numpy==1.26.4
packaging==24.2
pandas==1.2.5
paramiko==2.6.0
pillow==11.0.0
pyarrow==18.1.0
pycparser==2.22
pycryptodome==3.9.9
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.2
s3transfer==0.4.2
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==1.26.20
wget==3.2
zipp==3.21.0